Fix full-page reload on chat/admin navigation under static export
Next's client router falls back to a full page reload whenever it can't
fetch a route's RSC flight payload — which staticFrontend.ts was
unconditionally 404ing, even for static routes that next build already
wrote a real payload file for. Serving that file restores genuine soft
navigation for all static routes (admin pages, etc).

Chat's dynamic [chatId] segment can never have a pre-baked payload since
its value only exists at runtime, so it now uses a small router-free
client-side navigation primitive (lib/clientRouter.tsx) instead of
next/navigation, avoiding the remount entirely.
