+++ apps/frontend/app/chat/[chatId]/PageClient.tsx
@@ -1,50 +1,13 @@
 'use client'
-import { useContext, useEffect } from 'react'
-import { Chat } from '@/app/chat/components/Chat'
-import ChatPageContext from '@/app/chat/components/context'
-import toast from 'react-hot-toast'
-import { ChatHeader } from '../components/ChatHeader'
-import { useSWRJson } from '@/hooks/swr'
-import * as dto from '@/types/dto'
-import env from '@/lib/env'
-import { useUrlSegment } from '@/hooks/useUrlSegment'
+import { ChatSection } from '@/app/chat/components/ChatSection'
 
+// /chat/[chatId]'s static shell (see page.tsx: generateStaticParams only ever
+// bakes the '_' placeholder) renders the exact same ChatSection as bare
+// /chat — which chat, if any, it shows comes entirely from
+// ChatPageContext's urlChatId, not from this route match. See
+// ChatSection's own comment for why.
 const ChatPage = () => {
-  const {
-    state: { selectedConversation },
-    loadConversation,
-  } = useContext(ChatPageContext)
-
-  const chatId = useUrlSegment(1)
-  useEffect(() => {
-    // As ChatPageState.selectedConversation is shared by all chat pages, and kept
-    // when routing between them, we must ensure that it matches the
-    // page URL
-    if (selectedConversation?.id !== chatId) {
-      void loadConversation(chatId).catch(() => {
-        toast.error('Failed loading the chat')
-      })
-    }
-  }, [chatId, loadConversation, selectedConversation?.id])
-
-  useEffect(() => {
-    document.title = `${selectedConversation?.name ?? env.appDisplayName}`
-  }, [selectedConversation?.name])
-
-  const assistantId = selectedConversation?.assistantId
-  const { data: assistant } = useSWRJson<dto.UserAssistantWithSupportedMedia>(
-    `/api/me/assistants/${assistantId}`
-  )
-
-  if (selectedConversation?.id !== chatId || !assistant) {
-    return null
-  }
-  return (
-    <div className="flex-1 flex flex-col overflow-hidden">
-      <ChatHeader assistant={assistant} />
-      <Chat assistant={assistant} className="flex-1" supportedMedia={assistant.supportedMedia} />
-    </div>
-  )
+  return <ChatSection />
 }
 
 export default ChatPage
